MAINT: Moved newton solver to its own file
